Upgrade all libs (except aleph) to the latest versions
`io.prometheus/simpleclient` is now at end-of-life with version
`0.16.0`, they moved to a breaking `io.prometheus/client_java` 1.0.0
library in Sept'23. Upgrading `iapetos` to 0.16.0 version of
simpleclient to cut a tag before making breaking changes in this
library as well.

`aleph` based tests break when we upgrade from `0.4.6` to `0.6.4`, I
will fix them in a separate commit.